requirements.txt
elasticsearch
regex
docker
gdown
autocorrect
pandas
eel
nltk
python-twitter
flickrapi